backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.4.0",
    "channels-redis>=4.2.1",
    "channels[daphne]>=4.2.0",
    "django>=5.1.5",
    "django-celery-beat>=2.7.0",
    "django-cors-headers>=4.6.0",
    "django-ninja>=1.3.0",
    "django-storages[s3]>=1.14.4",
    "psycopg>=3.2.4",
    "psycopg-binary>=3.2.4",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "sentry-sdk[django]>=2.20.0",
]
